refactor setup_contract_test! macro
